+++ lib/moleculer/packets/discover.rb
@@ -5,6 +5,13 @@ module Moleculer
     ##
     # Represents a DISCOVER packet
     class Discover < Base
+      packet_attr :node_id, nil
+
+      def topic
+        return "#{super}.#{node_id}" if node_id
+
+        super
+      end
     end
   end
 end