+++ lib/moleculer/serializers/json.rb
@@ -11,7 +11,7 @@ module Moleculer
       end
 
       def serialize(message)
-        message.as_json.to_json
+        message.to_h.to_json
       end
 
       def deserialize(message)