+++ lib/moleculer/service/action.rb
@@ -46,7 +46,7 @@ module Moleculer
         @service.node
       end
 
-      def as_json
+      def to_h
         {
           name:    "#{@service.full_name}.#{name}",
           rawName: name,