+++ lib/moleculer/transporters/fake.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require_relative "base"
 
 module Moleculer
@@ -11,14 +13,18 @@ module Moleculer
         # in this case we want to use a class var as this needs to behave like a singleton to mimic how a global
         # transporter functions
         @@subscriptions ||= {} # rubocop:disable Style/ClassVars
+        @logger           = config.logger.get_child("[FAKE.TRANSPORTER]")
       end
 
       def subscribe(channel, &block)
-        @@subscriptions[channel] = block
+        @@subscriptions[channel] ||= []
+        @@subscriptions[channel] << block
       end
 
       def publish(packet)
-        @@subscriptions[packet.topic].call(packet)
+        @logger.debug "publishing packet to '#{packet.topic}'", packet.to_h
+        @logger.debug "processing #{@@subscriptions[packet.topic].length} callbacks for '#{packet.topic}'"
+        @@subscriptions[packet.topic].each { |c| c.call(packet) }
       end
 
       def start