+++ spec/spec_helper.rb
@@ -21,3 +21,5 @@ RSpec.configure do |config|
     c.syntax = :expect
   end
 end
+
+RSpec.shared_examples "base s"